Use a shared base class for the application instance views
